requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.26.0
python-dotenv>=1.0.0

# Similarity (optional)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# Development (optional)
pytest>=8.0
pytest-asyncio>=0.23
responses>=0.24
ruff>=0.1.0
